Refactored ketchup-text-input to kup-text-input. Updated KetchUP readme file with new rules.
