+++ packages/ketchup/src/components/kup-text-input/readme.md
@@ -1,4 +1,4 @@
-# ketchup-text-input
+# kup-text-input
