+++ packages/ketchup/src/index.html
@@ -70,11 +70,11 @@
             displayed-field="label"
             value-field="value"
         ></kup-radio>
-        <ketchup-text-input
+        <kup-text-input
             id="int1"
             is-clearable
             initial-value="ciao"
-        ></ketchup-text-input>
+        ></kup-text-input>
         <h4>Combo using portal</h4>
         <kup-combo is-clearable use-portal="true"></kup-combo>
         <h4>FLD 1</h4>
@@ -230,6 +230,7 @@
                             children: [],
                         },
                         showSubmit: true,
+                        submitLabel: 'Conferma 1',
                         valueField: 'obj'
                     };
                     fld_2.data = [
